% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BD.R
\name{BD.solve}
\alias{BD.solve}
\title{Birth Death Process}
\usage{
BD.solve(lambda, mu)
}
\arguments{
\item{lambda}{Array of birth rates.}

\item{mu}{Array of death rates.}
}
\value{
The discrete distribution of number of customers in system.
}
\description{
Solves the flow balance equation for a birth-death type queue with finite capacity n.
}
\examples{
mu = c(1, 2, 2)
lambda = c(3, 2, 1)
pn = BD.solve(lambda, mu)
print(pn)
 0.1176471 0.3529412 0.3529412 0.1764706
Where the first element corresponds to p0, the second to p1, and so on.
}
