% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closedJacksonNetworks.R
\name{cjn.summary}
\alias{cjn.summary}
\title{Closed Single-Server Jackson Network}
\usage{
cjn.summary(v, mu, N, K, convolution = TRUE)
}
\arguments{
\item{v}{solution to vR = v}

\item{mu}{Array of service rates}

\item{N}{number of customers in the network}

\item{K}{number of nodes}

\item{convolution}{If true, uses convolution algorithm,}
}
\value{
A summary of the measures of performance.
}
\description{
Closed Single-Server Jackson Network
}
\examples{
# given service rates mu = c(4, 1, 3)
number of customers in the system N = 3
number of nodes in the network K = 3
given service rates
mu = c(4, 1, 3)
N = 3
K = 3

First the using the convolution algorithm
q <- cjn.summary(v, mu, N, K, convolution = TRUE)

qt <- q$res
qt \%>\%
  gt() \%>\%
  tab_header(
    title = md("Measures of Performance for Closed Jackson Networks"))
marginal distribution
print(q$marginal_probabilities)

          [,1]       [,2]        [,3]
[1,] 0.68248175 0.04744526 0.788321168
[2,] 0.22554745 0.12481752 0.170802920
[3,] 0.07226277 0.29562044 0.035036496
[4,] 0.01970803 0.53211679 0.005839416

system throughput
print(q$system_throughput)

[1] 2.857664

using same inputs as above with the MVA Algorithm
q <- cjn.summary(v, mu, N, K, convolution = FALSE)

qt <- q$res
qt \%>\%
  gt() \%>\%
  tab_header(
     title = md("Measures of Performance for Closed Jackson Networks"))

print(q$system_throughput)
[1] 2.857664
}
