% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erlangloss.R
\name{erlangloss.summary}
\alias{erlangloss.summary}
\title{Erlang's Loss formula -- Erlang's B formula}
\usage{
erlangloss.summary(lambda, mu, c, plot_blocking = TRUE)
}
\arguments{
\item{lambda}{Arrival rate}

\item{mu}{Service Rate}

\item{c}{Number of servers/system capacity}

\item{plot_blocking}{Given c, plots the probability of blocking integers less than c.}
}
\value{
Returns a summary table of blocking probabilities, effective arrival rate per day,
number turned away, the average number of entities in system, and utilization for each server 1 to c, and a
plot of blocking probabilities.
}
\description{
The special case of the truncated queue M/M/c/K with K = c.
No line is allowed to form.
A stationary distribution known as Erlang's first formula.
}
\examples{
lambda <- 6
mu <- 3
c <- 4
erlangloss.summary(6,3,4)

$res
num_servers blocking lambda_eff turned_away    NS     U
         1    0.667      1.998       4.002 0.666 0.666
         2    0.400      3.600       2.400 1.200 0.600
         3    0.211      4.734       1.266 1.578 0.526
         4    0.095      5.430       0.570 1.810 0.452

$blocking_probabilities
[1] 0.667 0.400 0.211 0.095

$lambda_eff
[1] 1.998 3.600 4.734 5.430

$turned_away
[1] 4.002 2.400 1.266 0.570

$L
[1] 0.666 1.200 1.578 1.810

$Utilization
[1] 0.6660 0.6000 0.5260 0.4525

}
