% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finitepopulation.R
\name{finitepopulation.summary}
\alias{finitepopulation.summary}
\title{Finite Population Single Server Model (N >= 1)
A finite population queueing model consists of service requests generated
by fixed number of customers by a single server.}
\usage{
finitepopulation.summary(lambda, mu, N)
}
\arguments{
\item{lambda}{Arrival rate}

\item{mu}{Service Rate}

\item{N}{System Capacity}
}
\value{
Returns a data frame of performance measures.
}
\description{
Finite Population Single Server Model (N >= 1)
A finite population queueing model consists of service requests generated
by fixed number of customers by a single server.
}
\examples{
lambda <- 0.025
mu <- 1
N <- 4
q <- finitepopulation.summary(lambda, mu, N)

$pn
[1] 9.026213e-01 9.026213e-02
[3] 6.769660e-03 3.384830e-04
[5] 8.462074e-06

$res
Definition Result
I                       Idle time 0.9026
W             Mean time in system 1.0767
W_q            Mean time in queue 0.0767
L           Mean number in system 0.1049
L_q          Mean number in queue 0.0075
lambda_eff Effective arrival rate 0.0974

$p0
[1] 0.9026213

$W
[1] 1.076731

$W_q
[1] 0.07673145

$L
[1] 0.1048507

$L_q
[1] 0.007472012

$lambda_eff
[1] 0.09737873
}
