% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventOrientedBookkeeping.R
\name{gg1.summary}
\alias{gg1.summary}
\title{Title}
\usage{
gg1.summary(arrivals, service_times, interarrival = FALSE)
}
\arguments{
\item{arrivals}{Arrival time of customer}

\item{service_times}{Service time of customer}

\item{interarrival}{Boolean set to true, if arrival time is an interarrival time.}
}
\value{
Outputs Measures of Effectiveness for a G/G/1 queues.
}
\description{
Title
}
\examples{
arrival <- c(0, 2, 3, 6, 7, 8, 12, 14, 19, 20, 24, 26)
service <- c(1, 3, 6, 2, 1, 1, 1, 2, 5, 1, 1, 3)

inter <- c(1, 9, 6, 4, 7, 9, 5, 8, 4, 10, 6, 12, 6, 8,
          9, 5, 7, 8, 8, 7)
serv <- c(3, 7, 9, 9, 10, 4, 8, 5, 5, 3, 6, 3, 5, 4,
         9, 9, 8, 6, 8, 3)
res <- gg1.summary(arrival, service)
res2 <- gg1.summary(inter, serv, interarrival = TRUE)



}
