% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmc.R
\name{mmc.summary}
\alias{mmc.summary}
\title{M/M/c Markovian Model}
\usage{
mmc.summary(
  lambda,
  mu,
  c,
  tol,
  plot_pn = FALSE,
  plot_waitSys = FALSE,
  plot_waitQ = FALSE
)
}
\arguments{
\item{lambda}{Arrival Rate}

\item{mu}{Service Rate}

\item{c}{Number of Servers. c= 1 is a}

\item{plot_pn}{Boolean set to true, plots the distribution.}

\item{plot_waitSys}{Boolean set to true for a M/M/1, will plot the probability density function of the waiting time in system}

\item{plot_waitQ}{Boolean set to true for a M/M/c, plot the probability density function of the waiting time in queue}
}
\value{
Measures of Performance and Plots of waiting times.
}
\description{
M/M/c Markovian Model
}
\examples{
q <- mmc.summary(lambda = 4, mu = 4.4545, c = 1, plot_pn = TRUE,
plot_waitQ = TRUE,
plot_waitSys = TRUE)

"System is stable."
library(gt)
qt <- q$res
qt \%>\%
 gt() \%>\%
 tab_header(
   title = md("Measures of Performance for M/M/1 Queues"))


k = 2
prob_k_system <- (q$rho)^k
print(prob_k_system)
0.8063472

Markovian queueing models with parallel channels: M/M/c/∞/∞ queues.

q <- mmc.summary(lambda = 4, mu = 2, c = 3, plot_pn = TRUE)
 "System is stable."


library(gt)
qt <- q$res
qt \%>\%
 gt() \%>\%
 tab_header(
  title = md("Measures of Performance for M/M/c Queues"))
}
