% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openJacksonNetworks.R
\name{ojn.summary}
\alias{ojn.summary}
\title{Open Jackson Networks}
\usage{
ojn.summary(R, gamma, mu)
}
\arguments{
\item{R}{Routing matrix R.}

\item{gamma}{Mean arrival rate to each node.}

\item{mu}{Mean service rate of each node.}
}
\value{
Measures of performance.
}
\description{
Open Jackson Networks
}
\examples{
R = matrix(c(1/3, 1/4, 0.000, 1/4, 0.000, 1/6, 0.000,
1/3, 1/4, 0.000, 1/3, 0.000, 0.000, 0.000,
0.000, 0.000, 1/3, 1/3, 1/3, 0.000, 0.000,
1/3, 0.000, 1/3, 0.000, 1/3, 0.000, 0.000,
0.000, 0.000, 0.000, 4/5, 0.000, 0.000, 1/6,
1/6, 0.000, 1/6, 1/6, 1/6, 1/6, 0.000,
0.000, 1/6, 1/6, 1/6, 1/6, 0.000, 1/6),
nrow = 7, ncol = 7, byrow = TRUE)
gamma = c(0, 5, 0, 5, 0, 0, 0)
mu = c(85, 85, 120, 120, 70, 20, 20)

q <- ojn.summary(R, gamma, mu)

qt <- q$res
qt \%>\%
   gt() \%>\%
   tab_header(
      title = md("Measures of Performance for Open Jackson Networks"))


}
