% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closedJacksonNetworks.R
\name{solve.routing}
\alias{solve.routing}
\title{Solution to Routing Matrix}
\usage{
\method{solve}{routing}(R)
}
\arguments{
\item{R}{Matrix of routing probabilites}
}
\value{
Returns vR=v such that  ∑vi=1
}
\description{
Finds v given the routing matrix R, solution to vR = v.
}
\examples{
R = matrix(c(0, 0.75, 0.25,0.6666666667, 0, 0.3333333333,
1, 0, 0), nrow = 3, ncol = 3, byrow = TRUE)
v = solve.routing(R)
print(v)
          [,1]      [,2]      [,3]
[1,] 0.4444444 0.3333333 0.2222222
}
