% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solverouting.R
\name{solverouting}
\alias{solverouting}
\title{Solution to Routing Matrix}
\usage{
solverouting(R)
}
\arguments{
\item{R}{Matrix of routing probabilites}
}
\value{
Returns vR=v such that  sum of v_i = 1.
}
\description{
Finds v given the routing matrix R, solution to vR = v.
}
\examples{
R = matrix(c(0, 0.75, 0.25,0.6666666667, 0, 0.3333333333,
1, 0, 0), nrow = 3, ncol = 3, byrow = TRUE)
v = solverouting(R)
print(v)
          [,1]      [,2]      [,3]
[1,] 0.4444444 0.3333333 0.2222222

}
