% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closedJacksonNetworks.R
\name{cjn.summary}
\alias{cjn.summary}
\title{Closed Single-Server Jackson Network}
\usage{
cjn.summary(v, mu, N, K, convolution = TRUE)
}
\arguments{
\item{v}{Solution to vR = v}

\item{mu}{Array of service rates}

\item{N}{Number of customers in the network}

\item{K}{Number of nodes}

\item{convolution}{Boolean. If true, uses convolution algorithm,}
}
\value{
A summary of the measures of performance.
}
\description{
A system where N entities continuously travel inside the network. When convolution
is set to true, the function uses Buzen’s algorithm to solve for the summary statistics.
 Otherwise the function will use Mean Value Analysis.
}
\examples{
R = matrix(c(0, 0.75, 0.25,0.6666666667, 0, 0.3333333333,1, 0, 0), nrow = 3, ncol = 3, byrow = TRUE)
v = solve.routing(R)
mu = c(4, 1, 3)
N = 3
K = 3

Using the convolution algorithm
q.convolution <- cjn.summary(v, mu, N, K, convolution = TRUE)

q.convolution$marginal_probabilities
[,1]       [,2]        [,3]
[1,] 0.68248175 0.04744526 0.788321168
[2,] 0.22554745 0.12481752 0.170802920
[3,] 0.07226277 0.29562044 0.035036496
[4,] 0.01970803 0.53211679 0.005839416

q.convolution$res
Definition node 1 node 2 node 3
X            Node i throughput 1.2701 0.9526  0.635
X.1      Utilization at node i 0.3175 0.9526 0.2117
X.2 Probability node i is idle 0.6825 0.0474 0.7883
X.3        Mean time at node i 0.3379 2.4276 0.4069
L_i      Mean number at node i 0.4292 2.3124 0.2584

q.convolution$system_throughput
[1] 2.857664
q.convolution$throughput_i
[,1]      [,2]      [,3]
[1,] 1.270073 0.9525547 0.6350365

q.convolution$rho_i
[,1]      [,2]      [,3]
[1,] 0.3175182 0.9525547 0.2116788

q.convolution$W_i
[,1]     [,2]      [,3]
[1,] 0.337931 2.427586 0.4068966

q.convolution$L_i
[1] 0.4291971 2.3124088 0.2583942


Same inputs as above with the MVA Algorithm
q.MVA <- cjn.summary(v, mu, N, K, convolution = FALSE)

q.MVA$res
Definition node 1 node 2 node 3
lambda_i          Node i throughput 1.2701 0.9526  0.635
rho_i         Utilization at node i 0.3175 0.9526 0.2117
I_i      Probability node i is idle 0.6825 0.0474 0.7883
W_i             Mean time at node i 0.3379 2.4276 0.4069
L_i           Mean number at node i 0.4292 2.3124 0.2584

q.MVA$system_throughput
[1] 2.857664

q.MVA$throughput_i
[1] 1.2700730 0.9525547 0.6350365

q.MVA$rho_i
[1] 0.3175182 0.9525547 0.2116788

q.MVA$W_i
[1] 0.3379310 2.4275862 0.4068966

q.MVA$L_i
[1] 0.4291971 2.3124088 0.2583942
}
