% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmc.R
\name{mmc.summary}
\alias{mmc.summary}
\title{M/M/c Markovian Model}
\usage{
mmc.summary(
  lambda,
  mu,
  c,
  tol,
  plot_pn = FALSE,
  plot_waitSys = FALSE,
  plot_waitQ = FALSE
)
}
\arguments{
\item{lambda}{Arrival Rate}

\item{mu}{Service Rate}

\item{c}{Number of Servers.}

\item{plot_pn}{Boolean set to true, plots the distribution.}

\item{plot_waitSys}{Boolean set to true for a M/M/1, will plot the probability density function of the waiting time in system.}

\item{plot_waitQ}{Boolean set to true for a M/M/c, plot the probability density function of the waiting time in queue.}
}
\value{
Returns a data frame of performance measures, an array with the discrete distribution of
number of customers in system and plots of waiting times.
}
\description{
Given inputs arrival rate, service rate, and number of servers c=1 the function
mm1.summary() returns a data frame of performance measures for the queue and
an array with the discrete distribution of number of customers in system.
 Entering "plot_pn = TRUE", the function plots the distribution.
 For m/m/1 queues, by entering "plot_waitSys = TRUE" the function plots the probability
 density function of the waiting time in system, "plot_waitQ = TRUE" plots
 the probability density function of the waiting time in queue.
 The function stops finding additional probabilities when the sum of the pn is less than 1 - tol, the tolerance level.
}
\examples{
M/M/1 Markovian model
mm1.queue <- mmc.summary(lambda = 4, mu = 4.4545, c = 1,
tol = 0.01,
plot_pn = TRUE,
plot_waitQ = TRUE,
plot_waitSys = TRUE)

[1] "System is stable."

mm1.queue$pn
[1] 0.102031653 0.091621195 0.082272933 0.073878490 0.066340545 0.059571710
[7] 0.053493510 0.048035478 0.043134339 0.038733271 0.034781252 0.031232463
[13] 0.028045763 0.025184208 0.022614621 0.020307214 0.018235235 0.016374664
[19] 0.014703930 0.013203664 0.011856472 0.010646737 0.009560433 0.008584966
[25] 0.007709028 0.006922463 0.006216152 0.005581908 0.005012377 0.004500956
[31] 0.004041716 0.003629333 0.003259026 0.002926502 0.002627906 0.002359777
[37] 0.002119005 0.001902799 0.001708653 0.001534317 0.001377768 0.001237192
[43] 0.001110959
mm1.queue$res
Definition Result
rho           Utilization  0.898
I               Idle time  0.102
W     Mean time in system 2.2002
W_q    Mean time in queue 1.9757
L   Mean number in system 8.8009
L_q  Mean number in queue 7.9029
mm1.queue$rho
[1] 0.8979683
mm1.queue$p0
[1] 0.1020317
mm1.queue$W
[1] 2.20022
mm1.queue$W_q
[1] 1.975728
mm1.queue$L
[1] 8.80088
mm1.queue$L_q
[1] 7.90291

Markovian queueing models with parallel channels: M/M/c/∞/∞ queues

mmc.queue <- mmc.summary(lambda = 4, mu = 4.4545, c = 2, tol = 0.01, plot_pn = TRUE)
[1] "System is stable."
mmc.queue$pn
[1] 0.38027733 0.34147700 0.15331777 0.06883725 0.03090684 0.01387668
[7] 0.00623041

mmc.queue$res
Definition Result
U                                      Utilization  0.449
I                                        Idle time 0.3803
B                         Mean number busy servers  0.898
W                              Mean time in system 0.2812
W_q                             Mean time in queue 0.0567
L                            Mean number in system 1.1247
L_q                           Mean number in queue 0.2267
delayed Probability arriving entity waits in queue 0.2782

mmc.queue$rho
[1] 0.4489842
mmc.queue$p0
[1] 0.3802773
mmc.queue$W
[1] 0.2811728
mmc.queue$W_q
[1] 0.05668072
mmc.queue$L
[1] 1.124691
mmc.queue$L_q
[1] 0.2267229

}
