% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openJacksonNetworks.R
\name{ojn.summary}
\alias{ojn.summary}
\title{Open Jackson Networks}
\usage{
ojn.summary(R, gamma, mu)
}
\arguments{
\item{R}{Routing matrix R.}

\item{gamma}{Mean arrival rate for each node.}

\item{mu}{Mean service rate for each node.}
}
\value{
Measures of performance.
}
\description{
When convolution is set to true, the function uses Buzen’s algorithm to solve
for the summary statistics. Otherwise the function will use Mean Value Analysis.
In order to avoid rounding error, it is necessary to keep the probabilities in
the routing matrix as fractions if they are non-terminating, recurring decimal numbers.
}
\examples{
R = matrix(c(1/3, 1/4, 0.000, 1/4, 0.000, 1/6, 0.000,
1/3, 1/4, 0.000, 1/3, 0.000, 0.000, 0.000,
0.000, 0.000, 1/3, 1/3, 1/3, 0.000, 0.000,
1/3, 0.000, 1/3, 0.000, 1/3, 0.000, 0.000,
0.000, 0.000, 0.000, 4/5, 0.000, 0.000, 1/6,
1/6, 0.000, 1/6, 1/6, 1/6, 1/6, 0.000,
0.000, 1/6, 1/6, 1/6, 1/6, 0.000, 1/6),
nrow = 7, ncol = 7, byrow = TRUE)
gamma = c(0, 5, 0, 5, 0, 0, 0)
mu = c(85, 85, 120, 120, 70, 20, 20)

openjackson.queue <- ojn.summary(R, gamma, mu)

openjackson.queue$res
Definition  node 1  node 2  node 3   node 4  node 5  node 6  node 7
1              Node i throughput 79.7647 36.1412 64.9412 115.4118 64.9412 15.9529 12.9882
2          Utilization at node i  0.9384  0.4252  0.5412   0.9618  0.9277  0.7976  0.6494
3     Probability node i is idle  0.0616  0.5748  0.4588   0.0382  0.0723  0.2024  0.3506
4            Mean time at node i   0.191  0.0205  0.0182   0.2179  0.1977  0.2471  0.1426
5          Mean number at node i  15.236  0.7397  1.1795  25.1538 12.8372  3.9419  1.8523
6      Mean queue time at node i  0.1792  0.0087  0.0098   0.2096  0.1834  0.1971  0.0926
7 Mean number in queue at node i 14.2975  0.3145  0.6383  24.1921 11.9095  3.1442  1.2029

openjackson.queue$throughput_i
[,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]
[1,] 79.76471 36.14118 64.94118 115.4118 64.94118 15.95294 12.98824

openjackson.queue$rho_i
[,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
[1,] 0.9384083 0.4251903 0.5411765 0.9617647 0.9277311 0.7976471 0.6494118

openjackson.queue$idle
[,1]      [,2]      [,3]       [,4]       [,5]      [,6]      [,7]
[1,] 0.0615917 0.5748097 0.4588235 0.03823529 0.07226891 0.2023529 0.3505882

openjackson.queue$W_i
[,1]       [,2]       [,3]      [,4]      [,5]     [,6]      [,7]
[1,] 0.1910112 0.02046713 0.01816239 0.2179487 0.1976744 0.247093 0.1426174

openjackson.queue$L_i
[,1]      [,2]     [,3]     [,4]     [,5]    [,6]     [,7]
[1,] 15.23596 0.7397062 1.179487 25.15385 12.83721 3.94186 1.852349

openjackson.queue$Wq_i
[,1]        [,2]       [,3]      [,4]      [,5]     [,6]       [,7]
[1,] 0.1792465 0.008702426 0.00982906 0.2096154 0.1833887 0.197093 0.09261745

openjackson.queue$Lq_i
[,1]      [,2]      [,3]     [,4]     [,5]     [,6]     [,7]
[1,] 14.29755 0.3145159 0.6383107 24.19208 11.90948 3.144213 1.202937

openjackson.queue$W
[1] 6.094041

openjackson.queue$L
[1] 60.94041
}
